// Frontend/src/api/generated/model/indicatorDto.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TMB Control API
 * OpenAPI spec version: v1
 */
import type { IndicatorCategoryDto } from './indicatorCategoryDto';
import type { IndicatorStatusDto } from './indicatorStatusDto';
import type { IndicatorFrequencyDto } from './indicatorFrequencyDto';
import type { IndicatorHistoryDto } from './indicatorHistoryDto';

export interface IndicatorDto {
  id: string;
  /** @nullable */
  code: string | null;
  /** @nullable */
  name: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  value?: number | null;
  /** @nullable */
  unit: string | null;
  /** @nullable */
  date?: string | null;
  category: IndicatorCategoryDto;
  status: IndicatorStatusDto;
  frequency: IndicatorFrequencyDto;
  min?: number;
  max?: number;
  xLocation?: number;
  yLocation?: number;
  isErrorConfigured?: boolean;
  /** @nullable */
  minError?: number | null;
  /** @nullable */
  maxError?: number | null;
  isAlertConfigured?: boolean;
  /** @nullable */
  minAlert?: number | null;
  /** @nullable */
  maxAlert?: number | null;
  /** @nullable */
  history?: IndicatorHistoryDto[] | null;
}