// Frontend/src/api/generated/model/createIndicatorRequest.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TMB Control API
 * OpenAPI spec version: v1
 */

export interface CreateIndicatorRequest {
  /** @nullable */
  code: string | null;
  /** @nullable */
  name: string | null;
  /** @nullable */
  description: string | null;
  /** @nullable */
  unit: string | null;
  categoryId: number;
  frequencyId: number;
  min?: number;
  max?: number;
  /** @nullable */
  xLocation?: number | null;
  /** @nullable */
  yLocation?: number | null;
  isErrorConfigured?: boolean;
  /** @nullable */
  minError?: number | null;
  /** @nullable */
  maxError?: number | null;
  isAlertConfigured?: boolean;
  /** @nullable */
  minAlert?: number | null;
  /** @nullable */
  maxAlert?: number | null;
}