// Frontend/src/api/generated/model/indicatorsMasterDataDto.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TMB Control API
 * OpenAPI spec version: v1
 */
import type { IndicatorCategoryDto } from './indicatorCategoryDto';
import type { IndicatorFrequencyDto } from './indicatorFrequencyDto';

export interface IndicatorsMasterDataDto {
  /** @nullable */
  categories?: IndicatorCategoryDto[] | null;
  /** @nullable */
  frequencies?: IndicatorFrequencyDto[] | null;
}